Refactoring methods and improving the code
feat: 
- Refactoring the code on the controller and service classes
- Create a file dto to handle record classes
- Create a file exception to handle custom exceptions for the code

+++ src/main/java/com/minhaCarona/service/CarService.java
@@ -1,12 +1,11 @@
 package com.minhaCarona.service;
 
-
+import com.minhaCarona.exceptions.VehicleNotFoundException;
 import com.minhaCarona.model.Cars;
-import com.minhaCarona.model.CarsData;
+import com.minhaCarona.dto.CarsData;
 import com.minhaCarona.repository.CarsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -16,42 +15,48 @@ public class CarService {
     @Autowired
     private final CarsRepository carsRepository;
 
-    public CarService(CarsRepository carsRepository) {
+    public CarService ( CarsRepository carsRepository ) {
         this.carsRepository = carsRepository;
     }
 
-    public Cars registerVehicle(CarsData vehicleRecord) {
-        return carsRepository.save(new Cars(vehicleRecord));
+    public Cars registerVehicle ( CarsData vehicleRecord ) {
+        return carsRepository.save(new Cars( vehicleRecord ));
     }
 
     public List<Cars> findCars() {
         return carsRepository.findAll();
     }
 
-    public void deleteById(Long carId) {
+    public void deleteById ( Long carId ) {
+        Optional<Cars> carFinder = carsRepository.findById( carId );
+
+        if (carFinder.isEmpty()) {
+            throw new VehicleNotFoundException("Vehicle not found with ID: " + carId);
+        }
+
         carsRepository.deleteById(carId);
     }
 
-    public Cars editVehicle(Long carId, CarsData vehicleRecord) {
+    public Cars editVehicle ( Long carId, CarsData vehicleRecord ) {
 
-        try {
-            Cars existingVehicle = carsRepository.findById(carId).orElseThrow(() -> new RuntimeException("Vehicle not found" + carId));
+        return carsRepository.findById(carId)
+                .map(existingVehicle -> {
+                    existingVehicle.setCarModel(vehicleRecord.carModel());
+                    existingVehicle.setNumberOfSeats(vehicleRecord.numberOfSeats());
+                    existingVehicle.setPlate(vehicleRecord.plate());
+                    return carsRepository.save(existingVehicle);
+                })
+                .orElseThrow(() -> new VehicleNotFoundException("Vehicle not found with ID: " + carId));
 
-            existingVehicle.setCarModel(vehicleRecord.carModel());
-            existingVehicle.setPlate(vehicleRecord.plate());
-            existingVehicle.setNumberOfSeats(vehicleRecord.numberOfSeats());
-            return carsRepository.save(existingVehicle);
-
-        } catch (RuntimeException e) {
-            throw new RuntimeException("Error while editing vehicle: " + e.getMessage(), e);
-        }
     }
 
-    public Optional<Cars> findCarByName(String carModel) {
-        try {
-            return carsRepository.findByCarModelContainingIgnoreCase(carModel);
-        } catch (RuntimeException e) {
-            throw new RuntimeException("Error finding car");
+    public List < Cars > findCarByName ( String carModel ) {
+
+        List <Cars> listOfCars = carsRepository.findCarModel(carModel);
+        if (listOfCars.isEmpty()) {
+            throw new VehicleNotFoundException("Vehicle " + carModel + " not found");
         }
+
+        return listOfCars;
     }
 }
