+++ src/main/java/com/minhaCarona/service/CarpoolService.java
@@ -1,7 +1,9 @@
 package com.minhaCarona.service;
 
+import com.minhaCarona.exceptions.CarpoolNotFoundException;
+import com.minhaCarona.exceptions.VehicleNotFoundException;
 import com.minhaCarona.model.Carpool;
-import com.minhaCarona.model.CarpoolData;
+import com.minhaCarona.dto.CarpoolData;
 import com.minhaCarona.model.Cars;
 import com.minhaCarona.repository.CarpoolRepository;
 import com.minhaCarona.repository.CarsRepository;
@@ -9,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CarpoolService {
@@ -24,8 +27,9 @@ public class CarpoolService {
     }
 
     public Carpool registerCarpool(CarpoolData carpoolData) {
+
         Cars car = carsRepository.findById(carpoolData.carId())
-                .orElseThrow(() -> new RuntimeException("Car not found"));
+                .orElseThrow(() -> new VehicleNotFoundException ("Vehicle not found with ID: " + carpoolData.carId()));
         return carpoolRepository.save(new Carpool(carpoolData, car));
     }
 
@@ -34,6 +38,12 @@ public class CarpoolService {
     }
 
     public void deleteCarpool(Long carpoolId) {
+        Optional <Carpool> carpoolFinder = carpoolRepository.findById(carpoolId);
+
+        if (carpoolFinder.isEmpty()) {
+            throw new CarpoolNotFoundException("Carpool not found with ID" + carpoolId);
+        }
+
         carpoolRepository.deleteById(carpoolId);
     }
 }
